Add files via upload
The web scraper project done in Flask for the Flipkart website is a program designed to automatically collect data from the Flipkart website and display it in a structured format for easy analysis. The program utilizes the Flask framework, a micro web framework written in Python, which allows for the creation of web applications.

The web scraper is designed to scrape Flipkart's product pages, including product titles, prices, ratings, and descriptions. The scraper uses BeautifulSoup, a Python library used to parse HTML and XML documents, to extract relevant data from the website.

The user interface is designed using HTML, CSS, and  Flask's Jinja2 templating engine used to render dynamic content. The interface includes a search bar and a table that displays the scraped data.

The program is deployed using a virtual environment, which isolates the program's dependencies from the system's dependencies. The program can be accessed through a web browser, and it can be easily modified to scrape other websites or to add new features. Overall, the Flipkart web scraper project is a useful tool for quickly gathering data from the Flipkart website and displaying it in a structured format.